third_party/WebKit/LayoutTests/http/tests/devtools/startup/dedicated-workers-list-expected.txt

CONSOLE MESSAGE: line 20: Received message from worker 1
CONSOLE MESSAGE: line 20: Received message from worker 2
Tests that dedicated workers created before worker inspection was enabled will be reported to the front-end. Bug 72020

Found worker
Found worker
Worker removed
Worker removed
Worker inspection enabled
